+++ src/lib/Contextmenu_Link_Entry.svelte
@@ -2,7 +2,7 @@
 	import {strip_start} from '@ryanatkn/belt/string.js';
 	import type {Snippet} from 'svelte';
 
-	import {get_contextmenu} from '$lib/contextmenu_state.svelte.js';
+	import {contextmenu_context} from '$lib/contextmenu_state.svelte.js';
 
 	const DEFAULT_LINK_ICON = '🔗';
 
@@ -14,7 +14,7 @@
 
 	const {href, icon = DEFAULT_LINK_ICON, children}: Props = $props();
 
-	const contextmenu = get_contextmenu();
+	const contextmenu = contextmenu_context.get();
 
 	// TODO move or upstream? rename? `print_url`?
 	const format_url = (url: string, host: string = location.host): string => {