+++ src/lib/Library_Footer.svelte
@@ -19,13 +19,13 @@
 </script>
 
 <footer class="box">
-	{#if children}{@render children()}{/if}
+	{@render children?.()}
 	<div class="logo box panel p_lg shadow_inset_xs">
-		{#if logo_header}{@render logo_header()}{/if}
+		{@render logo_header?.()}
 		<a href={pkg.repo_url} rel="me" title="source code on GitHub"
 			>{#if logo}{@render logo()}{:else}<Github_Logo />{/if}</a
 		>
-		{#if logo_footer}{@render logo_footer()}{/if}
+		{@render logo_footer?.()}
 	</div>
 	{#if root_url}
 		<div class="root_url">