+++ src/lib/Library_Menu.svelte
@@ -1,9 +1,8 @@
 <script lang="ts">
 	import {page} from '$app/stores';
 	import {slide} from 'svelte/transition';
-	import {base} from '$app/paths';
 
-	import type {Tome} from '$lib/tome.js';
+	import {to_tome_pathname, type Tome} from '$lib/tome.js';
 	import type {Snippet} from 'svelte';
 
 	interface Props {
@@ -29,12 +28,13 @@
 		<li class="category">
 			{#if children}{@render children(category)}{:else}<h6>{category}</h6>{/if}
 			<ul class="unstyled">
-				{#each tomes as item (item.slug)}
+				{#each tomes as item (item)}
+					{@const pathname = to_tome_pathname(item)}
 					<li role="none" transition:slide>
 						<a
 							class="menu_item ellipsis line_height_lg"
-							href="{base}/library/{item.slug}"
-							class:selected={item.pathname === $page.url.pathname}>{item.name}</a
+							href={pathname}
+							class:selected={pathname === $page.url.pathname}>{item.name}</a
 						>
 					</li>
 				{/each}
@@ -45,20 +45,20 @@
 
 <style>
 	.library_menu {
-		width: 100%;
-		max-width: var(--library_menu_width);
+		width: var(--library_menu_width);
+		min-width: var(--library_menu_width);
 	}
 
 	.category {
 		margin: var(--space_xl6) 0;
 	}
 
-	/* TODO should be a CSS class or variable, maybe should be the default?
+	/* TODO @many should be a CSS class or variable, maybe should be the default?
 	problem is it doesn't work on .bg, maybe needs a variant/modifier in the name? */
-	ul a:hover {
+	a:hover {
 		background-color: var(--bg_5);
 	}
-	ul a:is(:active, .selected) {
+	a:is(:active, .selected) {
 		background-color: var(--bg_7);
 	}
 </style>