+++ src/routes/library/introduction/+page.svelte
@@ -1,13 +1,13 @@
 <script lang="ts">
-	import Tome_Detail from '$lib/Tome_Detail.svelte';
-	import {get_tome} from '$lib/tome.js';
+	import Tome_Content from '$lib/Tome_Content.svelte';
+	import {get_tome_by_name} from '$lib/tome.js';
 
 	const LIBRARY_ITEM_NAME = 'introduction';
 
-	const tome = get_tome(LIBRARY_ITEM_NAME);
+	const tome = get_tome_by_name(LIBRARY_ITEM_NAME);
 </script>
 
-<Tome_Detail {tome}>
+<Tome_Content {tome}>
 	<section>
 		<p>
 			Fuz is a Svelte UI library that builds on my CSS framework <a
@@ -32,4 +32,4 @@
 			<a href="https://github.com/ryanatkn/fuz#readme">the readme</a>.
 		</p>
 	</section>
-</Tome_Detail>
+</Tome_Content>