+++ src/routes/library/logos/+page.svelte
@@ -2,8 +2,8 @@
 	// TODO @many why is this import needed? `Code` already imports it. Fails in dev with SSR enabled without it. Is there a Vite config option that would be better? I tried the combinations of `ssr.external/noExternal/external` with `@ryanatkn/fuz_code` and `prismjs`.
 	import 'prismjs';
 	import Code from '@ryanatkn/fuz_code/Code.svelte';
-	import {get_tome} from '$lib/tome.js';
-	import Tome_Detail from '$lib/Tome_Detail.svelte';
+	import {get_tome_by_name} from '$lib/tome.js';
+	import Tome_Content from '$lib/Tome_Content.svelte';
 	import Fuz_Logo from '$lib/Fuz_Logo.svelte';
 	import Github_Logo from '$lib/Github_Logo.svelte';
 	import Mdn_Logo from '$lib/Mdn_Logo.svelte';
@@ -15,10 +15,10 @@
 	import Spiderspace_Logo from '$lib/Spiderspace_Logo.svelte';
 
 	const LIBRARY_ITEM_NAME = 'logos';
-	const tome = get_tome(LIBRARY_ITEM_NAME);
+	const tome = get_tome_by_name(LIBRARY_ITEM_NAME);
 </script>
 
-<Tome_Detail {tome}>
+<Tome_Content {tome}>
 	<section>
 		<ul class="box text_align_center unstyled">
 			<li>
@@ -86,7 +86,7 @@
 			</li>
 		</ul>
 	</section>
-</Tome_Detail>
+</Tome_Content>
 
 <style>
 	li {